Pull QuerySpecContext out into GraphFixture object
