+++ l_space/src/test/scala/io/mediachain/QuerySpec.scala
@@ -8,106 +8,19 @@ import org.specs2.execute.{AsResult, Result}
 import org.specs2.specification.ForEach
 import org.apache.tinkerpop.gremlin.orientdb.OrientGraphFactory
 
-case class QueryObjects(
-  person: Person,
-  personCanonical: Canonical,
-  photoBlob: PhotoBlob,
-  photoBlobCanonical: Canonical,
-  modifiedPhotoBlob: PhotoBlob,
-  extraPhotoBlob: PhotoBlob,
-  extraPhotoBlobCanonical: Canonical
-  )
-
-case class QuerySpecContext(graph: Graph, q: QueryObjects)
-
 object QuerySpec extends
   Specification with
-  ForEach[QuerySpecContext] with
+  ForEach[GraphFixture.Context] with
   XorMatchers {
   import Traversals.{GremlinScalaImplicits, VertexImplicits}
-  object Util {
-    // guarantees returned string is different from input
-    // TODO: accept distance
-    def mutate(s: String): String = {
-      val idx = Random.nextInt(s.length)
-      val chars = ('a' to 'z').toSet
-      val replaced = s.charAt(idx)
-      val replacing = (chars - replaced).toVector(Random.nextInt(chars.size - 1))
-      s.updated(idx, replacing)
-    }
 
-    val stuffI = Random.shuffle(List("can of peas",
-      "wishbone", "pair of glasses", "spool of wire", "wrench", "baseball hat", "television", "food",
-      "wallet", "jar of pickles", "tea cup", "sketch pad", "towel", "game CD", "steak knife", "slipper",
-      "pants", "sand paper", "boom box", "plush unicorn")).toIterator
-    val foodI = Random.shuffle(List("Preserved Peaches", "Brussels Sprouts", "Bananas", "Lettuce Salad",
-      "Olives", "Broiled Ham", "Cigars", "Mixed Green Salad", "Oyster Bay Asparagus", "Roast Lamb, Mint Sauce",
-      "Lemonade", "Consomme en Tasse", "Liqueurs", "Iced Tea", "Canadian Club", "Radis", "Escarole Salad",
-      "Preserved figs", "Potatoes, baked", "Macedoine salad")).toIterator
-    def getPhotoBlob: PhotoBlob = {
-      val title = stuffI.next
-      val desc = foodI.next
-      // FIXME: randomize date
-      val date = "2016-02-22T19:04:13+00:00"
-      PhotoBlob(None, title, desc, date, None)
-    }
-
-    def getModifiedPhotoBlob: PhotoBlob = {
-      val b = getPhotoBlob
-      b.copy(description = mutate(b.description))
-    }
-
-    val bodhisattvasI = Random.shuffle(List("Avalokitesvara",
-      "Manjushri", "Samantabhadra", "Kshitigarbha", "Maitreya", "Mahasthamaprapta", "Ākāśagarbha")).toIterator
-    def getPerson: Person = Person.create(bodhisattvasI.next)
-
-    def setupTree(graph: Graph): QueryObjects = {
-      // add photo and canonical
-      val photoBlob = getPhotoBlob
-      val photoBlobV = graph + photoBlob
-      val photoBlobCanonical = Canonical.create
-      val canonicalV = graph + photoBlobCanonical
-      canonicalV --- DescribedBy --> photoBlobV
-
-      // add a revision to a photo
-      val modifiedBlob = getModifiedPhotoBlob
-      val modifiedBlobV = graph + modifiedBlob
-      photoBlobV --- ModifiedBy --> modifiedBlobV
-
-      // add an author for the photo
-      val person = getPerson
-      val personV = graph + person
-      val personCanonical = Canonical.create
-      val personCanonicalV = graph + personCanonical
-      personCanonicalV --- DescribedBy --> personV
-      photoBlobV --- AuthoredBy --> personCanonicalV
-
-      // add decoy objects that we shouldn't see in a subtree
-      val extraPhotoBlob = getPhotoBlob
-      val extraPhotoBlobV = graph + extraPhotoBlob
-      val extraPhotoBlobCanonical = Canonical.create
-      val extraPhotoBlobCanonicalV = graph + extraPhotoBlobCanonical
-      extraPhotoBlobCanonicalV --- DescribedBy --> extraPhotoBlobV
-      extraPhotoBlobV --- AuthoredBy --> personCanonicalV
-
-      QueryObjects(
-        personV.toCC[Person],
-        personCanonicalV.toCC[Canonical],
-        photoBlobV.toCC[PhotoBlob],
-        canonicalV.toCC[Canonical],
-        modifiedBlobV.toCC[PhotoBlob],
-        extraPhotoBlobV.toCC[PhotoBlob],
-        extraPhotoBlobCanonicalV.toCC[Canonical])
-    }
-  }
 
   // TODO: can you figure out how to abstract out the connection creation?
-  def foreach[R: AsResult](f: QuerySpecContext => R): Result = {
+  def foreach[R: AsResult](f: GraphFixture.Context => R): Result = {
 
     lazy val graph = new OrientGraphFactory(s"memory:test-${math.random}").getNoTx()
     try {
-      val queryObjects = Util.setupTree(graph)
-      AsResult(f(QuerySpecContext(graph, queryObjects)))
+      AsResult(f(GraphFixture.Context(graph)))
     } finally {
       graph.database().drop()
     }
@@ -125,74 +38,74 @@ object QuerySpec extends
   """
 
   // TESTS BELOW
-  def findsPhoto = { context: QuerySpecContext =>
-    val queriedCanonical = Query.findPhotoBlob(context.graph, context.q.photoBlob)
+  def findsPhoto = { context: GraphFixture.Context =>
+    val queriedCanonical = Query.findPhotoBlob(context.graph, context.objects.photoBlob)
 
     queriedCanonical must beRightXor({ (c: Canonical) =>
-      c.canonicalID == context.q.photoBlobCanonical.canonicalID
+      c.canonicalID == context.objects.photoBlobCanonical.canonicalID
     })
   }
 
-  def findsTree = { context: QuerySpecContext =>
-    val tree = Query.findTreeForCanonical(context.graph, context.q.photoBlobCanonical)
+  def findsTree = { context: GraphFixture.Context =>
+    val tree = Query.findTreeForCanonical(context.graph, context.objects.photoBlobCanonical)
     tree.foreach(t => println(t.V.toList()))
     tree must beRightXor { (g: Graph) =>
       // Canonical itself
-      (context.q.photoBlobCanonical.id.flatMap(id => g.V(id).headOption) aka "canonical" must beSome) and
+      (context.objects.photoBlobCanonical.id.flatMap(id => g.V(id).headOption) aka "canonical" must beSome) and
       // PhotoBlob
-      (context.q.photoBlob.id.flatMap(id => g.V(id).headOption) aka "describing photoblob" must beSome) and
+      (context.objects.photoBlob.id.flatMap(id => g.V(id).headOption) aka "describing photoblob" must beSome) and
       // Modifying PhotoBlob
-      (context.q.modifiedPhotoBlob.id.flatMap(id => g.V(id).headOption) aka "modifying photoblob" must beSome) and
+      (context.objects.modifiedPhotoBlob.id.flatMap(id => g.V(id).headOption) aka "modifying photoblob" must beSome) and
       // Person
-      (context.q.person.id.flatMap(id => g.V(id).headOption) aka "person" must beSome) and
+      (context.objects.person.id.flatMap(id => g.V(id).headOption) aka "person" must beSome) and
       // Person canonical
-      (context.q.personCanonical.id.flatMap(id => g.V(id).headOption) aka "person canonical" must beSome) and
+      (context.objects.personCanonical.id.flatMap(id => g.V(id).headOption) aka "person canonical" must beSome) and
       // Another photoblob by same author
-      (context.q.extraPhotoBlob.id.flatMap(id => g.V(id).headOption) aka "extra photoblob" must beNone) and
+      (context.objects.extraPhotoBlob.id.flatMap(id => g.V(id).headOption) aka "extra photoblob" must beNone) and
       // Another photoblob's canonical
-      (context.q.extraPhotoBlobCanonical.id.flatMap(id => g.V(id).headOption) aka "extra canonical" must beNone)
+      (context.objects.extraPhotoBlobCanonical.id.flatMap(id => g.V(id).headOption) aka "extra canonical" must beNone)
     }
   }
 
-  def findsPerson = { context: QuerySpecContext =>
-    val queriedCanonical = Query.findPerson(context.graph, context.q.person)
+  def findsPerson = { context: GraphFixture.Context =>
+    val queriedCanonical = Query.findPerson(context.graph, context.objects.person)
 
     queriedCanonical must beRightXor { (person: Canonical) =>
-      person.canonicalID == context.q.personCanonical.canonicalID &&
+      person.canonicalID == context.objects.personCanonical.canonicalID &&
         person.id.isDefined
     }
   }
 
-  def findsAuthor = { context: QuerySpecContext =>
+  def findsAuthor = { context: GraphFixture.Context =>
     val queriedAuthor =
-      Query.findAuthorForBlob(context.graph, context.q.photoBlob)
+      Query.findAuthorForBlob(context.graph, context.objects.photoBlob)
 
     queriedAuthor must beRightXor { (c: Canonical) =>
-      c.canonicalID == context.q.personCanonical.canonicalID
+      c.canonicalID == context.objects.personCanonical.canonicalID
     }
   }
 
-  def findsWorks = { context: QuerySpecContext =>
-    val queriedWorks = Query.findWorks(context.graph, context.q.person)
+  def findsWorks = { context: GraphFixture.Context =>
+    val queriedWorks = Query.findWorks(context.graph, context.objects.person)
 
     queriedWorks must beRightXor { (s: List[Canonical]) =>
-      s.contains(context.q.photoBlobCanonical)
+      s.contains(context.objects.photoBlobCanonical)
     }
   }
 
-  def doesNotFindPhoto = { context: QuerySpecContext =>
-   val queryBlob = context.q.photoBlob.copy(
-      description = Util.mutate(context.q.photoBlob.description))
+  def doesNotFindPhoto = { context: GraphFixture.Context =>
+   val queryBlob = context.objects.photoBlob.copy(
+      description = GraphFixture.Util.mutate(context.objects.photoBlob.description))
 
     val queriedPhoto = Query.findPhotoBlob(context.graph, queryBlob)
 
     queriedPhoto must beLeftXor()
   }
 
-  def findsCanonicalForModifiedBlob = { context: QuerySpecContext =>
+  def findsCanonicalForModifiedBlob = { context: GraphFixture.Context =>
 
-    val parentCanonical = Query.findPhotoBlob(context.graph, context.q.photoBlob)
-    val childCanonical = Query.findCanonicalForBlob(context.graph, context.q.modifiedPhotoBlob)
+    val parentCanonical = Query.findPhotoBlob(context.graph, context.objects.photoBlob)
+    val childCanonical = Query.findCanonicalForBlob(context.graph, context.objects.modifiedPhotoBlob)
     (childCanonical must beRightXor()) and
       (childCanonical must_== parentCanonical)
   }
